requirements.txt
numpy==1.19.5
pandas==1.1.4
Pillow==8.0.1
foolbox==3.3.1
tensorflow==2.4.1
scikit-learn==0.20.3
tqdm==4.62.3